+++ apps/api/apps/root/src/root.module.ts
@@ -1,19 +1,21 @@
 import { Module } from '@nestjs/common';
-import { AppController } from './rest/app.controller';
-import { AppService } from './app.service';
 import { ConfigModule } from '@nestjs/config';
 import RootConfiguration from '../../../config/configuration';
 import { join } from 'path';
+import { MongooseServiceModule } from './mongoose.module';
+import { RestServiceModule } from './rest.module';
 
 @Module({
   imports: [
     ConfigModule.forRoot({
-      load: [RootConfiguration],
-      isGlobal: true,
+      load       : [RootConfiguration],
+      isGlobal   : true,
       envFilePath: join(__dirname, '..', '..', '..', '.env'),
     }),
+    MongooseServiceModule,
+    RestServiceModule,
   ],
-  controllers: [AppController],
-  providers: [AppService],
+  controllers: [],
+  providers  : [],
 })
-export class AppModule {}
+export class RootModule {}