+++ packages/types/src/mongo/pipes.ts
@@ -1,0 +1,5 @@
+export interface UrlQuery {
+  filter: string;
+  projection: string;
+  options: string;
+}