+++ packages/util/src/mongo/plugins/soft-delete.plugin.ts
@@ -1,0 +1,5 @@
+import { Schema, Document } from 'mongoose';
+
+export function softDeletePlugin(schema: Schema<Document>) {
+  console.log(schema);
+}