+++ apps/api/CHANGELOG.md
@@ -3,9 +3,9 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.0.1-next.1](https://personal-github.com/Tushar1998/ultimate-monorepo/compare/v0.0.1-next.0...v0.0.1-next.1) (2022-07-10)
+## [0.0.1-next.1](https://personal-github.com/tushar1998/ultimate-monorepo/compare/v0.0.1-next.0...v0.0.1-next.1) (2022-07-10)
 
-**Note:** Version bump only for package @ultimate/api
+**Note:** Version bump only for package @tushar1998/api
 
 
 
@@ -13,4 +13,4 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## 0.0.1-next.0 (2022-07-10)
 
-**Note:** Version bump only for package @ultimate/api
+**Note:** Version bump only for package @tushar1998/api